+++ src/components/display/index.ts
@@ -31,11 +31,19 @@ export * from "./skeleton-table"
 export * from "./skeleton-form"
 export {
   SmartCard as InfoCard,
+  SmartCard,
+  type SmartCardClassNames,
   type SmartCardClassNames as InfoCardClassNames,
+  type SmartCardDensity,
   type SmartCardDensity as InfoCardDensity,
+  type SmartCardOrientation,
   type SmartCardOrientation as InfoCardOrientation,
+  type SmartCardProps,
   type SmartCardProps as InfoCardProps,
+  type SmartCardRenderContext,
   type SmartCardRenderContext as InfoCardRenderContext,
+  type SmartCardSize,
   type SmartCardSize as InfoCardSize,
+  type SmartCardVariant,
   type SmartCardVariant as InfoCardVariant,
 } from "./smart-card"